Bugfixing (3)

+++ construction.cuh
@@ -8,6 +8,6 @@
 
 __global__ void build_quadtree_device(const U8ArraySoa& soa, Node *g_nodes, int tree_height, int n_rows, int n_cols, double detail_threshold);
 
-__host__ void build_quadtree_host(Node *quadtree_nodes, int depth, int height, int detail_threshold);
+__host__ void build_quadtree_host(Node *quadtree_nodes, int depth, int height, double detail_threshold);
 
 #endif//ACA_CUDA_CONSTRUCTION_CUH
